Create separate tool classes
