+++ crewai_tools/tools/patronus_eval_tool/patronus_local_evaluator_tool.py
@@ -1,0 +1,89 @@
+import os
+import json
+import requests
+import warnings
+from typing import Any, List, Dict, Optional, Type
+from crewai.tools import BaseTool
+from pydantic import BaseModel, Field
+from patronus import Client
+
+
+class FixedLocalEvaluatorToolSchema(BaseModel):
+    evaluated_model_input: Dict = Field(
+        ..., description="The agent's task description in simple text"
+    )
+    evaluated_model_output: Dict = Field(
+        ..., description="The agent's output of the task"
+    )
+    evaluated_model_retrieved_context: Dict = Field(
+        ..., description="The agent's context"
+    )
+    evaluated_model_gold_answer: Dict = Field(
+        ..., description="The agent's gold answer only if available"
+    )
+    evaluator: str = Field(..., description="The registered local evaluator")
+
+
+class PatronusLocalEvaluatorTool(BaseTool):
+    name: str = "Patronus Local Evaluator Tool"
+    evaluator: str = "The registered local evaluator"
+    evaluated_model_gold_answer: str = "The agent's gold answer"
+    description: str = (
+        "This tool is used to evaluate the model input and output using custom function evaluators."
+    )
+    client: Any = None
+    args_schema: Type[BaseModel] = FixedLocalEvaluatorToolSchema
+
+    class Config:
+        arbitrary_types_allowed = True
+
+    def __init__(self, evaluator: str, evaluated_model_gold_answer: str, **kwargs: Any):
+        super().__init__(**kwargs)
+        self.client = Client()
+        if evaluator:
+            self.evaluator = evaluator
+            self.evaluated_model_gold_answer = evaluated_model_gold_answer
+            self.description = f"This tool calls the Patronus Evaluation API that takes an additional argument in addition to the following new argument:\n evaluators={evaluator}, evaluated_model_gold_answer={evaluated_model_gold_answer}"
+            self._generate_description()
+            print(
+                f"Updating judge evaluator, gold_answer to: {self.evaluator}, {self.evaluated_model_gold_answer}"
+            )
+
+    def _run(
+        self,
+        **kwargs: Any,
+    ) -> Any:
+        evaluated_model_input = kwargs.get("evaluated_model_input")
+        evaluated_model_output = kwargs.get("evaluated_model_output")
+        evaluated_model_retrieved_context = kwargs.get(
+            "evaluated_model_retrieved_context"
+        )
+        evaluated_model_gold_answer = self.evaluated_model_gold_answer
+        evaluator = self.evaluator
+
+        result = self.client.evaluate(
+            evaluator=evaluator,
+            evaluated_model_input=(
+                evaluated_model_input
+                if isinstance(evaluated_model_input, str)
+                else evaluated_model_input.get("description")
+            ),
+            evaluated_model_output=(
+                evaluated_model_output
+                if isinstance(evaluated_model_output, str)
+                else evaluated_model_output.get("description")
+            ),
+            evaluated_model_retrieved_context=(
+                evaluated_model_retrieved_context
+                if isinstance(evaluated_model_retrieved_context, str)
+                else evaluated_model_retrieved_context.get("description")
+            ),
+            evaluated_model_gold_answer=(
+                evaluated_model_gold_answer
+                if isinstance(evaluated_model_gold_answer, str)
+                else evaluated_model_gold_answer.get("description")
+            ),
+            tags={},
+        )
+        output = f"Evaluation result: {result.pass_}, Explanation: {result.explanation}"
+        return output
